+++ DesignSystem/Sources/DesignSystem/Util/Color+.swift
@@ -1,5 +1,30 @@
 import SwiftUI
 
+extension Color {
+    public init?(hex: String) {
+        let r, g, b, a: CGFloat
+        if hex.hasPrefix("#") {
+            let start = hex.index(hex.startIndex, offsetBy: 1)
+            let hexColor = String(hex[start...])
+            
+            if hexColor.count == 6 {
+                let scanner = Scanner(string: hexColor)
+                var hexNumber: UInt64 = 0
+                
+                if scanner.scanHexInt64(&hexNumber) {
+                    r = CGFloat((hexNumber & 0xff0000) >> 16) / 255
+                    g = CGFloat((hexNumber & 0x00ff00) >> 8) / 255
+                    b = CGFloat(hexNumber & 0x0000ff) / 255
+                    a = 1.0
+                    self.init(red: r, green: g, blue: b, opacity: a)
+                    return
+                }
+            }
+        }
+        return nil
+    }
+}
+
 #if os(iOS)
 extension Color {
     // Common
