+++ DesignSystem/Sources/DesignSystem/Util/UIApplication.swift
@@ -1,10 +1,3 @@
-//
-//  File.swift
-//  DesignSystem
-//
-//  Created by 송형욱 on 1/7/25.
-//
-
 import Foundation
 import UIKit
